add pretrain setting to resnet backbones

+++ lib/losses.py
@@ -144,7 +144,7 @@ class CrossEntropyLoss(nn.Module):
         n_classes = pred.shape[1]
         if self.use_sigmoid:
             if n_classes == 1:
-                loss = F.binary_cross_entropy_with_logits(pred, label.view(-1, 1), reduction='none')
+                loss = F.binary_cross_entropy_with_logits(pred, label.view(-1, 1).float(), reduction='none')
                 return loss.sum() * self.loss_weight
             else:
                 tar_one_hot = utils.one_hot_embedding(label, n_classes+1)
